+++ client/waatea2_client/lib/models/showavailabilitydetail_model.dart
@@ -1,8 +1,12 @@
+import 'package:waatea2_client/models/user_model.dart';
+import 'package:waatea2_client/models/userprofile_model.dart';
+
 class ShowAvailabilityDetailModel {
   final int pk;
   final String name;
   final String email;
   final int level;
+  final UserProfileModel playerProfile;
   final String mobilephone;
   final int abonnement;
   int state;
@@ -13,6 +17,7 @@ class ShowAvailabilityDetailModel {
       required this.name,
       required this.email,
       required this.level,
+      required this.playerProfile,
       required this.mobilephone,
       required this.state,
       required this.abonnement,
@@ -24,6 +29,7 @@ class ShowAvailabilityDetailModel {
         name: json['name'],
         email: json['email'],
         level: json['profile']['level'],
+        playerProfile: UserProfileModel.fromJson(json['profile']),
         mobilephone: json['mobile_phone'],
         abonnement: json['profile']['abonnement'],
         state: 0,