+++ client/waatea2_client/lib/screens/editplayerdetail.dart
@@ -1,10 +1,10 @@
 import 'dart:convert';
 
 import 'package:flutter/material.dart';
+import 'package:waatea2_client/models/classification_model.dart';
 import 'package:waatea2_client/models/user_model.dart';
 import 'package:waatea2_client/screens/home.dart';
 import '../globals.dart' as globals;
-// ignore: depend_on_referenced_packages
 import 'package:http/http.dart' as http;
 
 import '../helper.dart';
@@ -22,6 +22,8 @@ class _EditPlayerDetailState extends State<EditPlayerDetail> {
   bool _isPlaying = false;
   int _selectedLevel = 1;
   int _selectedAbonnement = 0;
+  ClassificationModel? _selectedClassification; // Initialize as null
+  List<ClassificationModel> classificationOptions = [];
 
   @override
   void initState() {
@@ -29,6 +31,42 @@ class _EditPlayerDetailState extends State<EditPlayerDetail> {
     _isPlaying = widget.user.profile.isPlaying;
     _selectedLevel = widget.user.profile.level;
     _selectedAbonnement = widget.user.profile.abonnement;
+    //_selectedClassification = widget.user.profile.classification;
+
+    // Fetch and populate classification options
+    fetchClassifications().then((classifications) {
+      setState(() {
+        classificationOptions = classifications;
+        if (widget.user.profile.classification != null) {
+          // If the classification is not empty, set it based on the user's profile
+          _selectedClassification = classificationOptions.firstWhere(
+            (classification) =>
+                classification.pk == widget.user.profile.classification!.pk,
+            // Set to null when no match is found
+          );
+        } else {
+          // If the classification is empty, set it to null
+          _selectedClassification = null;
+        }
+      });
+    });
+  }
+
+  Future<List<ClassificationModel>> fetchClassifications() async {
+    final response = await http.get(
+      Uri.parse(
+          '${globals.URL_PREFIX}/api/classifications/filter?club=${globals.clubId}'),
+      headers: {
+        'Authorization': 'Token ${globals.token}',
+      },
+    );
+
+    if (response.statusCode == 200) {
+      final List<dynamic> data = json.decode(response.body);
+      return data.map((item) => ClassificationModel.fromJson(item)).toList();
+    } else {
+      throw Exception('Failed to load classifications');
+    }
   }
 
   @override
@@ -80,6 +118,29 @@ class _EditPlayerDetailState extends State<EditPlayerDetail> {
               items: itemsLevel,
             ),
             const SizedBox(height: 16),
+            const Text('Classification'),
+            DropdownButton<ClassificationModel>(
+              value: _selectedClassification,
+              onChanged: (value) {
+                setState(() {
+                  _selectedClassification = value!;
+                });
+              },
+              items: [
+                // Add a default "Select Classification" item as the first item
+                DropdownMenuItem<ClassificationModel>(
+                  value: null,
+                  child: const Text('Select Classification'),
+                ),
+                ...classificationOptions.map((classification) {
+                  return DropdownMenuItem<ClassificationModel>(
+                    value: classification,
+                    child: Text(classification.name),
+                  );
+                }).toList(),
+              ],
+            ),
+            const SizedBox(height: 16),
             const Text('Select Abo'),
             DropdownButton<int>(
               value: _selectedAbonnement,
@@ -97,7 +158,8 @@ class _EditPlayerDetailState extends State<EditPlayerDetail> {
                 final Map<String, dynamic> body = {
                   'is_playing': _isPlaying,
                   'level': _selectedLevel,
-                  'abonnement': _selectedAbonnement
+                  'abonnement': _selectedAbonnement,
+                  'classification': _selectedClassification?.pk,
                 };
 
                 final http.Response response = await http.patch(