+++ client/waatea2_client/lib/widgets/showavailabilitydetail_row.dart
@@ -1,6 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:waatea2_client/helper.dart';
 import 'package:url_launcher/url_launcher.dart';
+import 'package:waatea2_client/models/user_model.dart';
+import 'package:waatea2_client/models/userprofile_model.dart';
 
 class ShowAvailabilityDetailRow extends StatefulWidget {
   final String name;
@@ -9,6 +11,7 @@ class ShowAvailabilityDetailRow extends StatefulWidget {
   final int state;
   final String updated;
   final String game;
+  final UserProfileModel player;
 
   const ShowAvailabilityDetailRow(
       {Key? key,
@@ -17,7 +20,8 @@ class ShowAvailabilityDetailRow extends StatefulWidget {
       required this.level,
       required this.state,
       required this.updated,
-      required this.game})
+      required this.game,
+      required this.player})
       : super(key: key);
 
   @override
@@ -63,6 +67,22 @@ class _ShowAvailabilityDetailRowState extends State<ShowAvailabilityDetailRow> {
                 ),
               ),
               stateIcon,
+              const SizedBox(width: 40),
+              Expanded(
+                  flex: 1,
+                  child: Column(
+                      crossAxisAlignment: CrossAxisAlignment.center,
+                      children: [
+                        Icon(
+                          widget.player.classification?.icon != null
+                              ? IconData(
+                                  int.parse(widget.player.classification!.icon,
+                                      radix: 16),
+                                  fontFamily: 'MaterialIcons',
+                                )
+                              : Icons.highlight_off,
+                        )
+                      ])),
               const SizedBox(width: 40),
               Expanded(
                   flex: 1,