+++ server/waateaapp/admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
-from .models import Club, Game, Team, Season, Availability, Training, Attendance, CurrentSeason, HistoricalGame
-from waatea_2.users.models import UserProfile
+from .models import Club, Game, Team, Season, Availability, Training, Attendance, CurrentSeason, HistoricalGame, Links
+from waatea_2.users.models import UserProfile, Level, Classification
 
 admin.site.register(Club)
 admin.site.register(Team)
@@ -8,7 +8,9 @@ admin.site.register(Season)
 admin.site.register(Availability)
 admin.site.register(Attendance)
 admin.site.register(CurrentSeason)
-
+admin.site.register(Links)
+admin.site.register(Classification)
+admin.site.register(Level)
 @admin.register(Training)
 class TrainingAdmin(admin.ModelAdmin):
     list_display = ["date", "season", "club"]